Refactor file structure
* Add lib folder and moved library files into that folder
* create helper folder in Tests
