+++ Tests/game_end_checker_test.rb
@@ -1,5 +1,5 @@
-require './board'
-require './game_end_checker'
+require_relative '../lib/board'
+require_relative '../lib/game_end_checker'
 
 require 'minitest/autorun'
