+++ Tests/helpers/input_to_string_io.rb
