+++ Tests/win_patterns_test.rb
@@ -1,5 +1,5 @@
-require './board'
-require './win_patterns'
+require_relative '../lib/board'
+require_relative '../lib/win_patterns'
 
 require 'minitest/autorun'
