+++ lib/board.rb
@@ -1,4 +1,4 @@
-require_relative 'board_row'
+require_relative '../lib/board_row'
 
 class Board
   include Enumerable
