+++ lib/board_row.rb
