+++ lib/win_patterns.rb
